Fix units

+++ app/src/main/java/com/draco/eeku/models/Preset.kt
@@ -3,5 +3,5 @@ package com.draco.eeku.models
 data class Preset(
     val id: String,
     val displayName: String,
-    val map: Map<Float, Float>
+    val map: Map<Int, Float>
 )
